Move DiffCacheKey to App\Support namespace
DiffCacheKey is a utility, not an Action. Move it out of Actions/ so
arch tests pass without ignoring exceptions.
