+++ src/app/Actions/BuildDiffContextAction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Actions;
 
+use App\Support\DiffCacheKey;
 use Illuminate\Support\Facades\Cache;
 
 final readonly class BuildDiffContextAction
