+++ src/app/Support/DiffCacheKey.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Actions;
+namespace App\Support;
 
 final class DiffCacheKey
 {
