+++ src/tests/Arch/ActionsTest.php
@@ -4,7 +4,7 @@
  * Actions layer conventions:
  * - All Actions must be final readonly classes
  * - All Actions must have a handle() method
- * - Actions must have the "Action" suffix (except DiffCacheKey which is a misplaced utility)
+ * - Actions must have the "Action" suffix
  * - Actions must not depend on Livewire
  * - Actions must not depend on Illuminate\Http
  */
@@ -12,20 +12,17 @@ arch('actions are final readonly classes')
     ->expect('App\Actions')
     ->classes()
     ->toBeFinal()
-    ->toBeReadonly()
-    ->ignoring('App\Actions\DiffCacheKey');
+    ->toBeReadonly();
 
 arch('actions have a handle method')
     ->expect('App\Actions')
     ->classes()
-    ->toHaveMethod('handle')
-    ->ignoring('App\Actions\DiffCacheKey');
+    ->toHaveMethod('handle');
 
 arch('actions have the Action suffix')
     ->expect('App\Actions')
     ->classes()
-    ->toHaveSuffix('Action')
-    ->ignoring('App\Actions\DiffCacheKey');
+    ->toHaveSuffix('Action');
 
 arch('actions do not depend on livewire')
     ->expect('App\Actions')
